add kd detector
