+++ configs/ld/ld_gflv1_r101_r34_fpn_coco_1x.py
@@ -17,5 +17,4 @@ model = dict(
         out_channels=256,
         start_level=1,
         add_extra_convs='on_output',
-        num_outs=5),
-)
+        num_outs=5))
